+++ cli/cmd/replicaset.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"errors"
 	"fmt"
-	"os"
 	"regexp"
 	"strings"
 
@@ -12,7 +11,6 @@ import (
 	"github.com/tarantool/tt/cli/cmdcontext"
 	"github.com/tarantool/tt/cli/connect"
 	"github.com/tarantool/tt/cli/connector"
-	"github.com/tarantool/tt/cli/modules"
 	"github.com/tarantool/tt/cli/replicaset"
 	replicasetcmd "github.com/tarantool/tt/cli/replicaset/cmd"
 	"github.com/tarantool/tt/cli/running"
@@ -51,7 +49,6 @@ var (
 
 	chosenReplicasetAliases []string
 	lsnTimeout              int
-	downgradeVersion        string
 
 	replicasetUriHelp = "  The URI can be specified in the following formats:\n" +
 		"  * [tcp://][username:password@][host:port]\n" +
@@ -61,21 +58,15 @@ var (
 
 // newUpgradeCmd creates a "replicaset upgrade" command.
 func newUpgradeCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetUpgradeModule, &cobra.Command{
 		Use: "upgrade (<APP_NAME> | <URI>) [flags]\n\n" +
 			replicasetUriHelp,
 		DisableFlagsInUseLine: true,
 		Short:                 "Upgrade tarantool cluster",
 		Long: "Upgrade tarantool cluster.\n\n" +
 			libconnect.EnvCredentialsHelp + "\n\n",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetUpgradeModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
 		Args: cobra.ExactArgs(1),
-	}
+	})
 
 	cmd.Flags().StringArrayVarP(&chosenReplicasetAliases, "replicaset", "r",
 		[]string{}, "specify the replicaset name(s) to upgrade")
@@ -90,43 +81,35 @@ func newUpgradeCmd() *cobra.Command {
 
 // newDowngradeCmd creates a "replicaset downgrade" command.
 func newDowngradeCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use: "downgrade (<APP_NAME> | <URI>) [flags]\n\n" +
+	validateVersion := func(i int) cobra.PositionalArgs {
+		return func(cmd *cobra.Command, args []string) error {
+			var versionPattern = regexp.MustCompile(`^\d+\.\d+\.\d+$`)
+			if args[i] == "" {
+				return errors.New("need to specify the version to downgrade " +
+					"use --version (-v) option")
+			} else if !versionPattern.MatchString(args[i]) {
+				return errors.New("--version (-v) must be in the format " +
+					"'x.x.x', where x is a number")
+			}
+			return nil
+		}
+	}
+
+	cmd := setupTtModuleCmd(internalReplicasetDowngradeModule, &cobra.Command{
+		Use: "downgrade (<APP_NAME> | <URI>) VERSION [flags]\n\n" +
 			replicasetUriHelp,
 		DisableFlagsInUseLine: true,
 		Short:                 "Downgrade tarantool cluster",
 		Long: "Downgrade tarantool cluster.\n\n" +
 			libconnect.EnvCredentialsHelp + "\n\n",
-		Run: func(cmd *cobra.Command, args []string) {
-			var versionPattern = regexp.MustCompile(`^\d+\.\d+\.\d+$`)
-			if downgradeVersion == "" {
-				err := errors.New("need to specify the version to downgrade " +
-					"use --version (-v) option")
-				util.HandleCmdErr(cmd, err)
-				os.Exit(1)
-			} else if !versionPattern.MatchString(downgradeVersion) {
-				err := errors.New("--version (-v) must be in the format " +
-					"'x.x.x', where x is a number")
-				util.HandleCmdErr(cmd, err)
-				os.Exit(1)
-			}
-
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetDowngradeModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
-		Args: cobra.ExactArgs(1),
-	}
+		Args: cobra.MatchAll(cobra.ExactArgs(2), validateVersion(1)),
+	})
 
 	cmd.Flags().StringArrayVarP(&chosenReplicasetAliases, "replicaset", "r",
 		[]string{}, "specify the replicaset name(s) to downgrade")
 
 	cmd.Flags().IntVarP(&lsnTimeout, "timeout", "t", 5,
 		"timeout for waiting the LSN synchronization (in seconds)")
-
-	cmd.Flags().StringVarP(&downgradeVersion, "version", "v", "",
-		"version to downgrade the schema to")
 
 	addOrchestratorFlags(cmd)
 	addTarantoolConnectFlags(cmd)
@@ -135,7 +118,7 @@ func newDowngradeCmd() *cobra.Command {
 
 // newStatusCmd creates a "replicaset status" command.
 func newStatusCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetStatusModule, &cobra.Command{
 		Use: "status [--cartridge|--config|--custom] [flags] " +
 			"(<APP_NAME> | <APP_NAME:INSTANCE_NAME> | <URI>)\n\n" +
 			replicasetUriHelp,
@@ -143,14 +126,8 @@ func newStatusCmd() *cobra.Command {
 		Short:                 "Show a replicaset status",
 		Long: "Show a replicaset status.\n\n" +
 			libconnect.EnvCredentialsHelp + "\n\n",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetStatusModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
 		Args: cobra.ExactArgs(1),
-	}
+	})
 
 	addOrchestratorFlags(cmd)
 	addTarantoolConnectFlags(cmd)
@@ -159,7 +136,7 @@ func newStatusCmd() *cobra.Command {
 
 // newPromoteCmd creates a "replicaset promote" command.
 func newPromoteCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetPromoteModule, &cobra.Command{
 		Use: "promote [--cartridge|--config|--custom] [-f] [--timeout secs] [flags] " +
 			"(<APP_NAME:INSTANCE_NAME> | <URI>)\n\n" +
 			replicasetUriHelp,
@@ -167,14 +144,8 @@ func newPromoteCmd() *cobra.Command {
 		Short:                 "Promote an instance",
 		Long: "Promote an instance.\n\n" +
 			libconnect.EnvCredentialsHelp + "\n\n",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetPromoteModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
 		Args: cobra.ExactArgs(1),
-	}
+	})
 
 	addOrchestratorFlags(cmd)
 	addTarantoolConnectFlags(cmd)
@@ -191,19 +162,13 @@ func newPromoteCmd() *cobra.Command {
 
 // newDemoteCmd creates a "replicaset demote" command.
 func newDemoteCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetDemoteModule, &cobra.Command{
 		Use:                   "demote [-f] [--timeout secs] [flags] <APP_NAME:INSTANCE_NAME>",
 		DisableFlagsInUseLine: true,
 		Short:                 "Demote an instance",
 		Long:                  "Demote an instance.",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetDemoteModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
-		Args: cobra.ExactArgs(1),
-	}
+		Args:                  cobra.ExactArgs(1),
+	})
 
 	addOrchestratorFlags(cmd)
 	cmd.Flags().BoolVarP(&replicasetForce, "force", "f", false,
@@ -215,19 +180,13 @@ func newDemoteCmd() *cobra.Command {
 
 // newExpelCmd creates a "replicaset expel" command.
 func newExpelCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetExpelModule, &cobra.Command{
 		Use: "expel [-f] [--cartridge|--config|--custom] [--timeout secs] " +
 			"<APP_NAME:INSTANCE_NAME>",
 		Short: "Expel an instance from a replicaset",
 		Long:  "Expel an instance from a replicaset.",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetExpelModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
-		Args: cobra.ExactArgs(1),
-	}
+		Args:  cobra.ExactArgs(1),
+	})
 
 	addOrchestratorFlags(cmd)
 	cmd.Flags().BoolVarP(&replicasetForce, "force", "f", false,
@@ -240,18 +199,12 @@ func newExpelCmd() *cobra.Command {
 
 // newBootstrapCmd creates a "replicaset bootstrap" command.
 func newBootstrapCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetBootstrapModule, &cobra.Command{
 		Use:   "bootstrap [--timeout secs] [flags] <APP_NAME|APP_NAME:INSTANCE_NAME>",
 		Short: "Bootstrap an application or instance",
 		Long:  "Bootstrap an application or instance.",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetBootstrapModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
-		Args: cobra.ExactArgs(1),
-	}
+		Args:  cobra.ExactArgs(1),
+	})
 
 	addOrchestratorFlags(cmd)
 	cmd.Flags().BoolVarP(&replicasetBootstrapVshard, "bootstrap-vshard", "", false,
@@ -268,7 +221,7 @@ func newBootstrapCmd() *cobra.Command {
 
 // newBootstrapVShardCmd creates a "vshard bootstrap" command.
 func newBootstrapVShardCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetBootstrapVShardModule, &cobra.Command{
 		Use: "bootstrap [--cartridge|--config|--custom] [--timeout secs] [flags] " +
 			"(<APP_NAME> | <APP_NAME:INSTANCE_NAME> | <URI>)\n\n" +
 			replicasetUriHelp,
@@ -276,14 +229,8 @@ func newBootstrapVShardCmd() *cobra.Command {
 		Short:                 "Bootstrap vshard in the cluster",
 		Long: "Bootstrap vshard in the cluster.\n\n" +
 			libconnect.EnvCredentialsHelp + "\n\n",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetBootstrapVShardModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
 		Args: cobra.ExactArgs(1),
-	}
+	})
 
 	addOrchestratorFlags(cmd)
 	addTarantoolConnectFlags(cmd)
@@ -308,18 +255,12 @@ func newVShardCmd() *cobra.Command {
 
 // newRebootstrapCmd creates a "replicaset rebootstrap" command.
 func newRebootstrapCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetRebootstrapModule, &cobra.Command{
 		Use:                   "rebootstrap <APP_NAME:INSTANCE_NAME>",
 		DisableFlagsInUseLine: true,
 		Short:                 "Re-bootstraps an instance",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetRebootstrapModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
-		Args: cobra.ExactArgs(1),
-	}
+		Args:                  cobra.ExactArgs(1),
+	})
 
 	cmd.Flags().BoolVarP(&rebootstrapConfirmed, "yes", "y", false,
 		"automatically confirm rebootstrap")
@@ -340,19 +281,13 @@ func newRolesCmd() *cobra.Command {
 
 // newRolesAddCmd creates a "replicaset roles add" command.
 func newRolesAddCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetRolesAddModule, &cobra.Command{
 		Use: "add [--cartridge|--config|--custom] [-f] [--timeout secs]" +
 			"<APP_NAME:INSTANCE_NAME> <ROLE_NAME> [flags]",
 		Short: "Adds a role for Cartridge and Tarantool 3 orchestrator",
 		Long:  "Adds a role for Cartridge and Tarantool 3 orchestrator",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetRolesAddModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
-		Args: cobra.ExactArgs(2),
-	}
+		Args:  cobra.ExactArgs(2),
+	})
 
 	cmd.Flags().StringVarP(&replicasetReplicasetName, "replicaset", "r", "",
 		"name of a target replicaset")
@@ -378,19 +313,13 @@ func newRolesAddCmd() *cobra.Command {
 
 // newRolesRemoveCmd creates a "replicaset roles remove" command.
 func newRolesRemoveCmd() *cobra.Command {
-	cmd := &cobra.Command{
+	cmd := setupTtModuleCmd(internalReplicasetRolesRemoveModule, &cobra.Command{
 		Use: "remove [--cartridge|--config|--custom] [-f] [--timeout secs]" +
 			"<APP_NAME:INSTANCE_NAME> <ROLE_NAME> [flags]",
 		Short: "Removes a role for Cartridge and Tarantool 3 orchestrator",
 		Long:  "Removes a role for Cartridge and Tarantool 3 orchestrator",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalReplicasetRolesRemoveModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
-		Args: cobra.ExactArgs(2),
-	}
+		Args:  cobra.ExactArgs(2),
+	})
 
 	cmd.Flags().StringVarP(&replicasetReplicasetName, "replicaset", "r", "",
 		"name of a target replicaset")
@@ -479,7 +408,7 @@ type replicasetCtx struct {
 }
 
 // replicasetFillCtx fills the replicaset command context.
-func replicasetFillCtx(cmdCtx *cmdcontext.CmdCtx, ctx *replicasetCtx, args []string,
+func replicasetFillCtx(cmdCtx *cmdcontext.CmdCtx, ctx *replicasetCtx, target string,
 	isRunningCtxRequired bool) error {
 	var err error
 	ctx.Orchestrator, err = getOrchestrator()
@@ -497,7 +426,8 @@ func replicasetFillCtx(cmdCtx *cmdcontext.CmdCtx, ctx *replicasetCtx, args []str
 	}
 	var connOpts connector.ConnectOpts
 
-	if err := running.FillCtx(cliOpts, cmdCtx, &ctx.RunningCtx, args, false); err == nil {
+	err = running.FillCtx(cliOpts, cmdCtx, &ctx.RunningCtx, []string{target}, false)
+	if err == nil {
 		ctx.IsApplication = true
 		if len(ctx.RunningCtx.Instances) == 1 {
 			if connectCtx.Username != "" || connectCtx.Password != "" {
@@ -511,7 +441,7 @@ func replicasetFillCtx(cmdCtx *cmdcontext.CmdCtx, ctx *replicasetCtx, args []str
 				connectCtx,
 			)
 			ctx.IsInstanceConnect = true
-			appName, instName, found := strings.Cut(args[0], string(running.InstanceDelimiter))
+			appName, instName, found := strings.Cut(target, string(running.InstanceDelimiter))
 			if found {
 				if instName != ctx.RunningCtx.Instances[0].InstName {
 					return fmt.Errorf("instance %q not found", instName)
@@ -554,7 +484,7 @@ func replicasetFillCtx(cmdCtx *cmdcontext.CmdCtx, ctx *replicasetCtx, args []str
 		if isRunningCtxRequired {
 			return err
 		}
-		connOpts, _, err = resolveConnectOpts(cmdCtx, cliOpts, &connectCtx, args)
+		connOpts, err = resolveConnectOpts(cmdCtx, cliOpts, &connectCtx, target)
 		if err != nil {
 			return err
 		}
@@ -576,7 +506,7 @@ func replicasetFillCtx(cmdCtx *cmdcontext.CmdCtx, ctx *replicasetCtx, args []str
 // internalReplicasetUpgradeModule is a "upgrade" command for the replicaset module.
 func internalReplicasetUpgradeModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, false); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], false); err != nil {
 		return err
 	}
 	if ctx.IsInstanceConnect {
@@ -592,7 +522,7 @@ func internalReplicasetUpgradeModule(cmdCtx *cmdcontext.CmdCtx, args []string) e
 		SslCiphers:  replicasetSslCiphers,
 	}
 	var connOpts connector.ConnectOpts
-	connOpts, _, _ = resolveConnectOpts(cmdCtx, cliOpts, &connectCtx, args)
+	connOpts, _ = resolveConnectOpts(cmdCtx, cliOpts, &connectCtx, args[0])
 
 	return replicasetcmd.Upgrade(replicasetcmd.DiscoveryCtx{
 		IsApplication: ctx.IsApplication,
@@ -608,7 +538,7 @@ func internalReplicasetUpgradeModule(cmdCtx *cmdcontext.CmdCtx, args []string) e
 // internalReplicasetDowngradeModule is a "upgrade" command for the replicaset module.
 func internalReplicasetDowngradeModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, false); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], false); err != nil {
 		return err
 	}
 	if ctx.IsInstanceConnect {
@@ -624,7 +554,7 @@ func internalReplicasetDowngradeModule(cmdCtx *cmdcontext.CmdCtx, args []string)
 		SslCiphers:  replicasetSslCiphers,
 	}
 	var connOpts connector.ConnectOpts
-	connOpts, _, _ = resolveConnectOpts(cmdCtx, cliOpts, &connectCtx, args)
+	connOpts, _ = resolveConnectOpts(cmdCtx, cliOpts, &connectCtx, args[0])
 
 	return replicasetcmd.Downgrade(replicasetcmd.DiscoveryCtx{
 		IsApplication: ctx.IsApplication,
@@ -634,14 +564,14 @@ func internalReplicasetDowngradeModule(cmdCtx *cmdcontext.CmdCtx, args []string)
 	}, replicasetcmd.DowngradeOpts{
 		ChosenReplicasetAliases: chosenReplicasetAliases,
 		Timeout:                 lsnTimeout,
-		DowngradeVersion:        downgradeVersion,
+		DowngradeVersion:        args[1],
 	}, connOpts)
 }
 
 // internalReplicasetPromoteModule is a "promote" command for the replicaset module.
 func internalReplicasetPromoteModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, false); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], false); err != nil {
 		return err
 	}
 	if !ctx.IsInstanceConnect {
@@ -671,7 +601,7 @@ func internalReplicasetPromoteModule(cmdCtx *cmdcontext.CmdCtx, args []string) e
 // internalReplicasetDemoteModule is a "demote" command for the replicaset module.
 func internalReplicasetDemoteModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, true); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], true); err != nil {
 		return err
 	}
 	if !ctx.IsApplication {
@@ -703,7 +633,7 @@ func internalReplicasetDemoteModule(cmdCtx *cmdcontext.CmdCtx, args []string) er
 // internalReplicasetStatusModule is a "status" command for the replicaset module.
 func internalReplicasetStatusModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, false); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], false); err != nil {
 		return err
 	}
 	if ctx.IsInstanceConnect {
@@ -723,7 +653,7 @@ func internalReplicasetExpelModule(cmdCtx *cmdcontext.CmdCtx, args []string) err
 		return fmt.Errorf("the command expects argument application_name:instance_name")
 	}
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, true); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], true); err != nil {
 		return err
 	}
 	if ctx.IsInstanceConnect {
@@ -750,7 +680,7 @@ func internalReplicasetExpelModule(cmdCtx *cmdcontext.CmdCtx, args []string) err
 // the "replicaset vshard" module.
 func internalReplicasetBootstrapVShardModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, false); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], false); err != nil {
 		return err
 	}
 	if ctx.IsInstanceConnect {
@@ -777,7 +707,7 @@ func internalReplicasetBootstrapModule(cmdCtx *cmdcontext.CmdCtx, args []string)
 	_, instName, found := strings.Cut(args[0], string(running.InstanceDelimiter))
 
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, true); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], true); err != nil {
 		return err
 	}
 	if ctx.IsInstanceConnect {
@@ -839,7 +769,7 @@ func internalReplicasetRebootstrapModule(cmdCtx *cmdcontext.CmdCtx, args []strin
 // internalReplicasetRolesAddModule is a "roles add" command for the replicaset module.
 func internalReplicasetRolesAddModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, false); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], false); err != nil {
 		return err
 	}
 	defer ctx.Conn.Close()
@@ -882,7 +812,7 @@ func internalReplicasetRolesAddModule(cmdCtx *cmdcontext.CmdCtx, args []string) 
 // internalReplicasetRolesRemoveModule is a "roles remove" command for the replicaset module.
 func internalReplicasetRolesRemoveModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	var ctx replicasetCtx
-	if err := replicasetFillCtx(cmdCtx, &ctx, args, false); err != nil {
+	if err := replicasetFillCtx(cmdCtx, &ctx, args[0], false); err != nil {
 		return err
 	}
 	defer ctx.Conn.Close()