+++ cli/cmd/rocks.go
@@ -3,26 +3,18 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 	"github.com/tarantool/tt/cli/cmdcontext"
-	"github.com/tarantool/tt/cli/modules"
 	"github.com/tarantool/tt/cli/rocks"
-	"github.com/tarantool/tt/cli/util"
 )
 
 // NewRocksCmd creates rocks command.
 func NewRocksCmd() *cobra.Command {
-	var rocksCmd = &cobra.Command{
+	var rocksCmd = setupTtModuleCmd(internalRocksModule, &cobra.Command{
 		Use:   "rocks",
 		Short: "LuaRocks package manager",
 		// Disabled all flags parsing on this commands leaf.
 		// LuaRocks will handle it self.
 		DisableFlagParsing: true,
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalRocksModule, args)
-			util.HandleCmdErr(cmd, err)
-		},
-	}
+	})
 
 	return rocksCmd
 }